requirements.txt
python-dotenv>=0.19.0
requests>=2.26.0
pandas>=1.3.0
python-linkedin-v2>=0.8.3
indeed-python>=0.0.6
beautifulsoup4>=4.9.3
spacy>=3.1.0
python-telegram-bot>=13.7
fastapi>=0.68.0
uvicorn>=0.15.0
pyresparser>=1.0.6
